Updated versions of eTrack scripts
New bdam.py script to calculate number of neighbouring atoms for each
atom in structure (using pipeline by Markus’ thesis) - with use of CCP4
program pdbcur; function finally calculates Bdamage for each atom in
structure, using all atoms within neighbouring distance of 14 Angstroms.

New bdamchange_vs_atomnum.py script to plot Bdamage change as function
of atom number with each dataset number on separate axis, AND to plot
Bdamage change and Bfactor change as function of atom number on same
plot.

Minor alternations in classholder_v2.py script to determine RNA base
types (A,G,C,U).. extending from previous DNA work (DA,DC,DG,DT).

Inclusion of previously written colrowsec_to_cartesian.py script to
convert unit cell coordinate system to cartesian xyz system, to
ultimately calculate basis vectors in cartesian space.

New script find_metricchange.py to calculate both the Bfactor and
Bdamage changes per atom between an initial and later dataset —> saves
results as atom attributes.

inputfile_ccp4jobs.txt altered now to reflect fact that the TRAP
datasets are now being processed (as opposed to 2BN3 insulin datasets).

Updating of map2voxelclasslist_nocartconversion.py script. Only
function densmap2class_readvoxels modified (which is not used in main
run of eTrack). This was performed to allow more flexibility to
ultimately just read in a density map —> with no attention to whether
map is density map or atom map.

New side script mapfile_hist.py was created to plot a kde plot of
electron density values split between the protein structure and solvent
contributions to the map file. This is not needed in the eTrack script,
it is merely an unpolished side script.

A new script was written called metricVSdensity_Multi.py to simply plot
a scatter plot of a density metric (mean,median,min,max) against an
atomic metric (Bfactor, Bdamage, etc), for each dataset individually,
to determine correlations between the two chosen metrics. There is a
choice to only include particular residue/atom/chain types here. Each
dataset number is coloured differently in the scatter plot.

The master script multifile_handle_GENERAL.py was modified to now read
in the TRAPRNA damage datasets. Additionally the post_processing
function was modified to now also read in the initial structure pdb
file as a list of atom objects (like the later datasets), and to then
calculate the number of surrounding atoms for this initial structure.
This number of surrounding atoms was then extended to each later
dataset (without a repeated calculation - which is very time
consuming), and then the Bdamage and Bdamage change calculated for each
atom in each of the later datasets. This part is currently commented
out for the TRAP processing since it takes a long time to run. It will
be uncommented out when needed for further analysis. Additionally the
graphanalysis function now includes further function calls to plot
graphs of Bfactor and Bdamage change against atom number.

In the PDBfile_manipulation.py script, the loading bar imported from
progbar.progress was now used in the for loop (which currently takes a
long time to run) in the location of all common atoms present in all
datasets —> basically so the user knows the program hasn’t crashed..

In the script res_formatter.py the script is now more generalised to
read in all the residue/base types present (without a predefined list)
and to order these alphabetically (with no preference towards amino
acids or nucleotides). This is a change to the previous system, in
which the DNA bases were listed after the ordered amino acid residue
types appearing in the structure.

Finally the script TRAPRNA_study.py is specialised to my investigation
of the TRAP-RNA complex. It contains a function to plot violin plots
for the TRA-RNA case study to determine whether there is any
heterogeneity between the spread of protein damage between the
RNA-bound and non-RNA protein TRAP macromolecules —> currently
considering the min,max,mean density metrics, as well as Bfactor. This
function only selects atoms within 4.5 Angstrom of the RNA binding
sites (with equivalent atoms included for the non RNA bound TRAP
molecule), to allow a more targeted analysis to the binding sites.
Furthermore there is a function densitychange_v_atomnum to plot mean
density change AND min density change as a function of atom number on
the same plot (as subplots), with all damage datasets appearing in a
different colour on each subplot. Another function called
TRAPringcompare_format outputs a multidimensional list of atoms in the
structure organised as follows: each atom in chain A and it’s
equivalent (same residue type and number, same atom type) for all the
other TRAP chains (21 others in total) form a list, and all these lists
become elements in a collective list. The function TRAPringcompare_plot
then reads this list in, and allows the user to specify which atom
types, residue types, residue numbers (to effectively locate one
specific atom in a chain by its identity) to then plot a series of line
plots (one subplot for each dataset) with chain number (1—>22) on the
x-axis and specified density metric (min, mean, or max) on the y-axis.